move trial durations to util/trial
